@@ -8,9 +8,9 @@
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
     "is-svg": "^4.2.2",
-    "react": "^17.0.2",
+    "react": "^18.3.1",
     "react-bootstrap": "^1.5.2",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.3.1",
     "react-player": "^2.16.0",
     "react-responsive-masonry": "^2.3.0",
     "react-router-bootstrap": "^0.25.0",
